Add Customer, Role, Permission Controllers

+++ Application/Services/EmployeeService.cs
@@ -24,34 +24,37 @@ public class EmployeeService : IEmployeeService
 
     public Employee? AddEmployee(User user, Employee employee)
     {
-        var returnedUser = userRepository.GetFirstOrDefault(x => x.UserName == user.UserName
-        && x.PasswordHash == user.PasswordHash);
+        //var returnedUser = userRepository.GetFirstOrDefault(x => x.UserName == user.UserName
+        //&& x.PasswordHash == user.PasswordHash);
 
-        if (returnedUser == null)
-        {
-            returnedUser = userRepository.Add(user);
-        }
+        //if (returnedUser == null)
+        //{
+        //    returnedUser = userRepository.Add(user);
+        //}
 
-        employee.UserId = returnedUser.Id;
+        //employee.UserId = returnedUser.Id;
 
         employee = employeeRepository.Add(employee);
+
+        employeeRepository.SaveChanges();
 
         return employee;
     }
 
     public async Task<Employee?> AddEmployeeAsync(User user, Employee employee)
     {
-        var returnedUser = await userRepository.GetFirstOrDefaultAsync(x => x.UserName == user.UserName
-        && x.PasswordHash == user.PasswordHash);
+        //var returnedUser = await userRepository.GetFirstOrDefaultAsync(x => x.UserName == user.UserName
+        //&& x.PasswordHash == user.PasswordHash);
 
-        if (returnedUser == null)
-        {
-            returnedUser = await userRepository.AddAsync(user);
-        }
+        //if (returnedUser == null)
+        //{
+        //    returnedUser = await userRepository.AddAsync(user);
+        //}
 
-        employee.UserId = returnedUser.Id;
+        //employee.UserId = returnedUser.Id;
 
         employee = await employeeRepository.AddAsync(employee);
+        await employeeRepository.SaveChangesAsync();
 
         return employee;
     }
@@ -61,8 +64,16 @@ public class EmployeeService : IEmployeeService
         var employee = employeeRepository.Remove(id);
         employeeRepository.SaveChanges();
 
-        userRepository.Remove(id);
+        if (employee == null)
+            return null;
+
+        var user = userRepository.Remove(id);
         userRepository.SaveChanges();
+
+        if (user == null)
+            return null;
+
+        employee.User = user;
 
         return employee;
     }
@@ -72,15 +83,73 @@ public class EmployeeService : IEmployeeService
         var employee = await employeeRepository.RemoveAsync(id);
         await employeeRepository.SaveChangesAsync();
 
-        await userRepository.RemoveAsync(id);
+        if (employee == null) 
+            return null;
+
+        var user = await userRepository.RemoveAsync(id);
         await userRepository.SaveChangesAsync();
 
+        if (user == null)
+            return null;
+
+        employee.User = user;
+
         return employee;
+    }
+
+    public IEnumerable<Employee>? GetAllEmployees()
+    {
+        var employees = employeeRepository.GetAll();
+
+        if (employees == null)
+            return null;
+
+        foreach (var employee in employees)
+        {
+            var user = userRepository.GetById(employee.UserId);
+
+            if (user == null)
+                continue;
+
+            employee.User = user;
+        }
+
+        return employees;
+    }
+
+    public async Task<IEnumerable<Employee>?> GetAllEmployeesAsync()
+    {
+        var employees = await employeeRepository.GetAllAsync();
+
+        if (employees == null)
+            return null;
+
+        foreach (var employee in employees)
+        {
+            var user = await userRepository.GetByIdAsync(employee.UserId);
+
+            if (user == null)
+                continue;
+
+            employee.User = user;
+        }
+
+        return employees;
     }
 
     public Employee? GetEmployeeById(Guid id)
     {
         var employee = employeeRepository.GetById(id);
+
+        if (employee == null)
+            return null;
+
+        var user = userRepository.GetById(id);
+
+        if (user == null)
+            return null;
+
+        employee.User = user;
 
         return employee;
     }
@@ -88,6 +157,16 @@ public class EmployeeService : IEmployeeService
     public async Task<Employee?> GetEmployeeByIdAsync(Guid id)
     {
         var employee = await employeeRepository.GetByIdAsync(id);
+
+        if (employee == null)
+            return null;
+
+        var user = await userRepository.GetByIdAsync(id);
+
+        if (user == null)
+            return null;
+
+        employee.User = user;
 
         return employee;
     }
@@ -141,6 +220,16 @@ public class EmployeeService : IEmployeeService
         var returnedEmployee = employeeRepository.Update(id, employee);
         employeeRepository.SaveChanges();
 
+        if (returnedEmployee == null)
+            return null;
+
+        var user = userRepository.GetById(id);
+
+        if (user is null)
+            return null;
+
+        returnedEmployee.User = user;
+
         return returnedEmployee;
     }
 
@@ -148,6 +237,16 @@ public class EmployeeService : IEmployeeService
     {
         var returnedEmployee = await employeeRepository.UpdateAsync(id, employee);
         await employeeRepository.SaveChangesAsync();
+
+        if (returnedEmployee == null)
+            return null;
+
+        var user = await userRepository.GetByIdAsync(id);
+
+        if (user is null)
+            return null;
+
+        returnedEmployee.User = user;
 
         return returnedEmployee;
     }
